Split into the different threading mechanisms
