+++ libcudacxx/include/cuda/std/__atomic/wait/polling.h
@@ -53,7 +53,7 @@ template <typename _Tp, typename _Sco>
 _CCCL_HOST_DEVICE void __atomic_try_wait_slow_fallback(
   _Tp const volatile* __a, __atomic_underlying_remove_cv_t<_Tp> __val, memory_order __order, _Sco)
 {
-  __libcpp_thread_poll_with_backoff(__atomic_poll_tester<_Tp, _Sco>(__a, __val, __order));
+  _CUDA_VSTD::__libcpp_thread_poll_with_backoff(__atomic_poll_tester<_Tp, _Sco>(__a, __val, __order));
 }
 
 _LIBCUDACXX_END_NAMESPACE_STD
