+++ libcudacxx/include/cuda/std/__atomic/wait/threading_support_cuda.h
@@ -1,0 +1,48 @@
+// -*- C++ -*-
+//===----------------------------------------------------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
+#define _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
+
+#include <cuda/std/detail/__config>
+
+#if defined(_CCCL_IMPLICIT_SYSTEM_HEADER_GCC)
+#  pragma GCC system_header
+#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_CLANG)
+#  pragma clang system_header
+#elif defined(_CCCL_IMPLICIT_SYSTEM_HEADER_MSVC)
+#  pragma system_header
+#endif // no system header
+
+#if !defined(_LIBCUDACXX_HAS_NO_THREADS) && defined(_LIBCUDACXX_HAS_THREAD_API_CUDA)
+
+#  include <cuda/std/chrono>
+#  include <cuda/std/climits>
+
+_CCCL_PUSH_MACROS
+
+_LIBCUDACXX_BEGIN_NAMESPACE_STD
+
+_LIBCUDACXX_HIDE_FROM_ABI void __libcpp_thread_yield() {}
+
+_LIBCUDACXX_HIDE_FROM_ABI void __libcpp_thread_sleep_for(_CUDA_VSTD::chrono::nanoseconds __ns)
+{
+  NV_IF_TARGET(NV_IS_DEVICE,
+               (auto const __step = __ns.count(); assert(__step < numeric_limits<unsigned>::max());
+                asm volatile("nanosleep.u32 %0;" ::"r"((unsigned) __step)
+                             :);))
+}
+
+_LIBCUDACXX_END_NAMESPACE_STD
+
+_CCCL_POP_MACROS
+
+#endif // !_LIBCUDACXX_HAS_NO_THREADS && _LIBCUDACXX_HAS_THREAD_API_CUDA
+
+#endif // _LIBCUDACXX___ATOMIC_WAIT_THREADING_SUPPORT_CUDA_H
